Move MultiNewArrayBasicBlockBuilder to GC Common plugin
Fixes #1205.
